requirements.txt
flask==3.1.2
flask-cors==6.0.2
flask-socketio==5.6.0
python-socketio==5.16.0
python-engineio==4.13.0
ace-framework==0.7.1
litellm==1.80.11
groq==1.0.0
python-dotenv==1.2.1
gunicorn==21.2.0
eventlet==0.35.2
